feat(core): Use JWTs in Invitation URLs
