+++ packages/cli/src/services/user.service.ts
@@ -1,25 +1,18 @@
-import { Container, Service } from 'typedi';
+import { Service } from 'typedi';
 import type { IUserSettings } from 'n8n-workflow';
-import { ApplicationError, ErrorReporterProxy as ErrorReporter } from 'n8n-workflow';
+import { ApplicationError } from 'n8n-workflow';
 
-import { type AssignableRole, User } from '@db/entities/User';
+import { type User } from '@db/entities/User';
 import { UserRepository } from '@db/repositories/user.repository';
 import type { PublicUser } from '@/Interfaces';
 import type { PostHogClient } from '@/posthog';
-import { Logger } from '@/Logger';
-import { UserManagementMailer } from '@/UserManagement/email';
-import { InternalHooks } from '@/InternalHooks';
-import { UrlService } from '@/services/url.service';
-import type { UserRequest } from '@/requests';
-import { InternalServerError } from '@/errors/response-errors/internal-server.error';
+import { InvitationService } from './invitation.service';
 
 @Service()
 export class UserService {
 	constructor(
-		private readonly logger: Logger,
 		private readonly userRepository: UserRepository,
-		private readonly mailer: UserManagementMailer,
-		private readonly urlService: UrlService,
+		private readonly invitationService: InvitationService,
 	) {}
 
 	async update(userId: string, data: Partial<User>) {
@@ -61,7 +54,10 @@ export class UserService {
 		}
 
 		if (options?.withInviteUrl && options?.inviterId && publicUser.isPending) {
-			publicUser = this.addInviteUrl(options.inviterId, publicUser);
+			publicUser.inviteAcceptUrl = this.invitationService.generateInvitationUrl(
+				options.inviterId,
+				user.id,
+			);
 		}
 
 		if (options?.posthog) {
@@ -73,17 +69,6 @@ export class UserService {
 		}
 
 		return publicUser;
-	}
-
-	private addInviteUrl(inviterId: string, invitee: PublicUser) {
-		const url = new URL(this.urlService.getInstanceBaseUrl());
-		url.pathname = '/signup';
-		url.searchParams.set('inviterId', inviterId);
-		url.searchParams.set('inviteeId', invitee.id);
-
-		invitee.inviteAcceptUrl = url.toString();
-
-		return invitee;
 	}
 
 	private async addFeatureFlags(publicUser: PublicUser, posthog: PostHogClient) {
@@ -101,118 +86,5 @@ export class UserService {
 		});
 
 		return await Promise.race([fetchPromise, timeoutPromise]);
-	}
-
-	private async sendEmails(
-		owner: User,
-		toInviteUsers: { [key: string]: string },
-		role: AssignableRole,
-	) {
-		const domain = this.urlService.getInstanceBaseUrl();
-
-		return await Promise.all(
-			Object.entries(toInviteUsers).map(async ([email, id]) => {
-				const inviteAcceptUrl = `${domain}/signup?inviterId=${owner.id}&inviteeId=${id}`;
-				const invitedUser: UserRequest.InviteResponse = {
-					user: {
-						id,
-						email,
-						inviteAcceptUrl,
-						emailSent: false,
-					},
-					error: '',
-				};
-
-				try {
-					const result = await this.mailer.invite({
-						email,
-						inviteAcceptUrl,
-						domain,
-					});
-					if (result.emailSent) {
-						invitedUser.user.emailSent = true;
-						delete invitedUser.user?.inviteAcceptUrl;
-						void Container.get(InternalHooks).onUserTransactionalEmail({
-							user_id: id,
-							message_type: 'New user invite',
-							public_api: false,
-						});
-					}
-
-					void Container.get(InternalHooks).onUserInvite({
-						user: owner,
-						target_user_id: Object.values(toInviteUsers),
-						public_api: false,
-						email_sent: result.emailSent,
-						invitee_role: role, // same role for all invited users
-					});
-				} catch (e) {
-					if (e instanceof Error) {
-						void Container.get(InternalHooks).onEmailFailed({
-							user: owner,
-							message_type: 'New user invite',
-							public_api: false,
-						});
-						this.logger.error('Failed to send email', {
-							userId: owner.id,
-							inviteAcceptUrl,
-							domain,
-							email,
-						});
-						invitedUser.error = e.message;
-					}
-				}
-
-				return invitedUser;
-			}),
-		);
-	}
-
-	async inviteUsers(owner: User, attributes: Array<{ email: string; role: AssignableRole }>) {
-		const emails = attributes.map(({ email }) => email);
-
-		const existingUsers = await this.userRepository.findManyByEmail(emails);
-
-		const existUsersEmails = existingUsers.map((user) => user.email);
-
-		const toCreateUsers = attributes.filter(({ email }) => !existUsersEmails.includes(email));
-
-		const pendingUsersToInvite = existingUsers.filter((email) => email.isPending);
-
-		const createdUsers = new Map<string, string>();
-
-		this.logger.debug(
-			toCreateUsers.length > 1
-				? `Creating ${toCreateUsers.length} user shells...`
-				: 'Creating 1 user shell...',
-		);
-
-		try {
-			await this.getManager().transaction(
-				async (transactionManager) =>
-					await Promise.all(
-						toCreateUsers.map(async ({ email, role }) => {
-							const newUser = transactionManager.create(User, { email, role });
-							const savedUser = await transactionManager.save<User>(newUser);
-							createdUsers.set(email, savedUser.id);
-							return savedUser;
-						}),
-					),
-			);
-		} catch (error) {
-			ErrorReporter.error(error);
-			this.logger.error('Failed to create user shells', { userShells: createdUsers });
-			throw new InternalServerError('An error occurred during user creation');
-		}
-
-		pendingUsersToInvite.forEach(({ email, id }) => createdUsers.set(email, id));
-
-		const usersInvited = await this.sendEmails(
-			owner,
-			Object.fromEntries(createdUsers),
-			attributes[0].role, // same role for all invited users
-		);
-
-		return { usersInvited, usersCreated: toCreateUsers.map(({ email }) => email) };
 	}
 }
